+++ app/destinations/destinations.php
@@ -17,7 +17,7 @@
 
 	The Initial Developer of the Original Code is
 	Mark J Crane <markjcrane@fusionpbx.com>
-	Portions created by the Initial Developer are Copyright (C) 2008-2018
+	Portions created by the Initial Developer are Copyright (C) 2008-2019
 	the Initial Developer. All Rights Reserved.
 
 	Contributor(s):
@@ -65,69 +65,58 @@
 	}
 
 //get variables used to control the order
-	$order_by = check_str($_GET["order_by"]);
-	$order = check_str($_GET["order"]);
+	$order_by = $_GET["order_by"];
+	$order = $_GET["order"];
 
 //set the type
-	if ($_GET['type'] == 'inbound') {
-		$destination_type = 'inbound';
-	}
-	elseif ($_GET['type'] == 'outbound') {
-		$destination_type = 'outbound';
-	}
-	elseif ($_GET['type'] == 'local') {
-		$destination_type = 'local';
-	}
-	else {
-		$destination_type = 'inbound';
+	switch ($_GET['type']) {
+		case 'inbound': $destination_type = 'inbound'; break;
+		case 'outbound': $destination_type = 'outbound'; break;
+		case 'local': $destination_type = 'local'; break;
+		default: $destination_type = 'inbound';
 	}
 
 //add the search term
-	$search = strtolower(check_str($_GET["search"]));
+	$search = strtolower($_GET["search"]);
 	if (strlen($search) > 0) {
 		$sql_search = " (";
-		$sql_search .= "lower(destination_type) like '%".$search."%' ";
-		$sql_search .= "or lower(destination_number) like '%".$search."%' ";
-		$sql_search .= "or lower(destination_context) like '%".$search."%' ";
-		$sql_search .= "or lower(destination_accountcode) like '%".$search."%' ";
+		$sql_search .= "lower(destination_type) like :search ";
+		$sql_search .= "or lower(destination_number) like :search ";
+		$sql_search .= "or lower(destination_context) like :search ";
+		$sql_search .= "or lower(destination_accountcode) like :search ";
 		if (permission_exists('outbound_caller_id_select')) {
-			$sql_search .= "or lower(destination_caller_id_name) like '%".$search."%' ";
-			$sql_search .= "or destination_caller_id_number like '%".$search."%' ";
+			$sql_search .= "or lower(destination_caller_id_name) like :search ";
+			$sql_search .= "or destination_caller_id_number like :search ";
 		}
-		$sql_search .= "or lower(destination_enabled) like '%".$search."%' ";
-		$sql_search .= "or lower(destination_description) like '%".$search."%' ";
+		$sql_search .= "or lower(destination_enabled) like :search ";
+		$sql_search .= "or lower(destination_description) like :search ";
+		$sql_search .= "or lower(destination_data) like :search ";
 		$sql_search .= ") ";
 	}
 
-//additional includes
-	require_once "resources/header.php";
-	require_once "resources/paging.php";
-
-//prepare to page the results
-	$sql = "select count(destination_uuid) as num_rows from v_destinations ";
-	$sql .= "where destination_type = '".$destination_type."' ";
+//common sql where
+	$sql_where = "where destination_type = :destination_type ";
 	if ($_GET['show'] == "all" && permission_exists('destination_all')) {
 		//show all
-	} else {
-		$sql .= "and (domain_uuid = '".$domain_uuid."' or domain_uuid is null) ";
+	}
+	else {
+		$sql_where .= "and (domain_uuid = :domain_uuid or domain_uuid is null) ";
+		$parameters['domain_uuid'] = $domain_uuid;
 	}
 	if (isset($sql_search)) {
-			$sql .= "and ".$sql_search;
+		$sql_where .= "and ".$sql_search;
+		$parameters['search'] = '%'.$search.'%';
 	}
-	if (strlen($order_by)> 0) { $sql .= "order by $order_by $order "; }
-	$prep_statement = $db->prepare($sql);
-	if ($prep_statement) {
-		$prep_statement->execute();
-		$row = $prep_statement->fetch(PDO::FETCH_ASSOC);
-		if ($row['num_rows'] > 0) {
-			$num_rows = $row['num_rows'];
-		}
-		else {
-			$num_rows = '0';
-		}
-	}
+	$parameters['destination_type'] = $destination_type;
 
 //prepare to page the results
+	$sql = "select count(destination_uuid) from v_destinations ";
+	$sql .= $sql_where;
+	$database = new database;
+	$num_rows = $database->select($sql, $parameters, 'column');
+
+//prepare to page the results
+	require_once "resources/paging.php";
 	$rows_per_page = ($_SESSION['domain']['paging']['numeric'] != '') ? $_SESSION['domain']['paging']['numeric'] : 50;
 	$param = "&search=".escape($search);
 	if ($_GET['show'] == "all" && permission_exists('destination_all')) {
@@ -135,27 +124,17 @@
 	}
 	$page = $_GET['page'];
 	if (strlen($page) == 0) { $page = 0; $_GET['page'] = 0; }
+	list($paging_controls_mini, $rows_per_page, $var_3) = paging($num_rows, $param, $rows_per_page, true);
 	list($paging_controls, $rows_per_page, $var3) = paging($num_rows, $param, $rows_per_page);
 	$offset = $rows_per_page * $page;
 
 //get the list
-	$sql = "select * from v_destinations ";
-	$sql .= "where destination_type = '".$destination_type."' ";
-	if ($_GET['show'] == "all" && permission_exists('destination_all')) {
-		//show all
-	} else {
-		$sql .= "and (domain_uuid = '".$domain_uuid."' or domain_uuid is null) ";
-	}
-	if (isset($sql_search)) {
-		$sql .= "and ".$sql_search;
-	}
-	$sql .= "and destination_type = '".$destination_type."' ";
-	if (strlen($order_by)> 0) { $sql .= "order by $order_by $order "; }
-	$sql .= "limit $rows_per_page offset $offset ";
-	$prep_statement = $db->prepare(check_sql($sql));
-	$prep_statement->execute();
-	$destinations = $prep_statement->fetchAll(PDO::FETCH_NAMED);
-	unset ($prep_statement, $sql);
+	$sql = str_replace('count(destination_uuid)', '*', $sql);
+	$sql .= order_by($order_by, $order);
+	$sql .= limit_offset($rows_per_page, $offset);
+	$database = new database;
+	$destinations = $database->select($sql, $parameters, 'all');
+	unset($parameters);
 
 //get the destination select list
 	$destination = new destinations;
@@ -186,6 +165,9 @@
 	$c = 0;
 	$row_style["0"] = "row_style0";
 	$row_style["1"] = "row_style1";
+
+//include the header
+	require_once "resources/header.php";
 
 //define the checkbox_toggle function
 	echo "<script type=\"text/javascript\">\n";
@@ -230,6 +212,9 @@
 
 	echo "				<input type='text' class='txt' style='width: 150px; margin-left: 15px;' name='search' id='search' value='".escape($search)."'>\n";
 	echo "				<input type='submit' class='btn' name='submit' value='".$text['button-search']."'>\n";
+	if ($paging_controls_mini != '') {
+		echo "			<span style='margin-left: 15px;'>".$paging_controls_mini."</span>\n";
+	}
 	echo "			</td>\n";
 	echo "		</form>\n";
 	echo "	</tr>\n";
@@ -251,8 +236,12 @@
 	}
 	echo th_order_by('destination_type', $text['label-destination_type'], $order_by, $order, $param);
 	echo th_order_by('destination_number', $text['label-destination_number'], $order_by, $order, $param);
-	echo  "<th>". $text['label-detail_action']."</th>";
-	echo th_order_by('destination_context', $text['label-destination_context'], $order_by, $order, $param);
+	if (!$_GET['show'] == "all") {
+		echo  "<th>". $text['label-detail_action']."</th>";
+	}
+	if (permission_exists("destination_context")) {
+		echo th_order_by('destination_context', $text['label-destination_context'], $order_by, $order, $param);
+	}
 	if (permission_exists('outbound_caller_id_select')) {
 		echo th_order_by('destination_caller_id_name', $text['label-destination_caller_id_name'], $order_by, $order, $param);
 		echo th_order_by('destination_caller_id_number', $text['label-destination_caller_id_number'], $order_by, $order, $param);
@@ -268,7 +257,7 @@
 	}
 	echo "	</td>\n";
 	echo "<tr>\n";
-	if (is_array($destinations)) {
+	if (is_array($destinations) && @sizeof($destinations) != 0) {
 		$x = 0;
 		foreach($destinations as $row) {
 			$action_name = action_name($destination_array, $row['destination_app'].':'.$row['destination_data']);
@@ -284,7 +273,7 @@
 			echo "	</td>\n";
 			if ($_GET['show'] == "all" && permission_exists('destination_all')) {
 				if (strlen($_SESSION['domains'][$row['domain_uuid']]['domain_name']) > 0) {
-					$domain = escape($_SESSION['domains'][$row['domain_uuid']]['domain_name']);
+					$domain = $_SESSION['domains'][$row['domain_uuid']]['domain_name'];
 				}
 				else {
 					$domain = $text['label-global'];
@@ -293,9 +282,13 @@
 			}
 			echo "	<td valign='top' class='".$row_style[$c]."'>".escape($row['destination_type'])."&nbsp;</td>\n";
 			echo "	<td valign='top' class='".$row_style[$c]."'>".escape(format_phone($row['destination_number']))."&nbsp;</td>\n";
-			echo "	<td valign='top' class='".$row_style[$c]."'>".escape($action_name)."&nbsp;</td>\n";
+			if (!$_GET['show'] == "all") {
+				echo "	<td valign='top' class='".$row_style[$c]."'>".$action_name."&nbsp;</td>\n";
+			}
 			//echo "	<td valign='top' class='".$row_style[$c]."'>".$row['destination_number_regex']."&nbsp;</td>\n";
-			echo "	<td valign='top' class='".$row_style[$c]."'>".escape($row['destination_context'])."&nbsp;</td>\n";
+			if (permission_exists("destination_context")) {
+				echo "	<td valign='top' class='".$row_style[$c]."'>".escape($row['destination_context'])."&nbsp;</td>\n";
+			}
 			//echo "	<td valign='top' class='".$row_style[$c]."'>".escape($row['fax_uuid'])."&nbsp;</td>\n";
 			if (permission_exists('outbound_caller_id_select')) {
 				echo "	<td valign='top' class='".$row_style[$c]."'>".escape($row['destination_caller_id_name'])."&nbsp;</td>\n";
@@ -313,7 +306,7 @@
 				echo "<a href='destination_edit.php?id=".escape($row['destination_uuid'])."' alt='".$text['button-edit']."'>$v_link_label_edit</a>";
 			}
 			if (permission_exists('destination_delete')) {
-				echo "<button type='submit' class='btn btn-default list_control_icon' name=\"destinations[$x][action]\" alt='".$text['button-delete']."' value='delete'><span class='glyphicon glyphicon-remove'></span></button>";
+				echo "<button type='submit' class='btn btn-default list_control_icon' name=\"destinations[$x][action]\" alt='".$text['button-delete']."' value='delete'><span class='fas fa-minus'></span></button>";
 			}
 			echo "	</td>\n";
 			echo "</tr>\n";
